+++ tests/rig/src/utils/mod.rs
@@ -16,7 +16,7 @@ use std::ops::Add;
 use std::path::PathBuf;
 pub use zksync_os_api::helpers::*;
 use zksync_os_interface::types::BlockOutput;
-use zksync_os_tests_common::zksync_tx::ZKsyncL1Tx;
+use zksync_os_tests_common::zksync_tx::l1_tx::ZKsyncL1Tx;
 use zksync_os_tests_common::zksync_tx::ZKsyncTxEnvelope;
 
 pub mod tx_encoding;
